+++ src/main/java/de/eldoria/shepard/contexts/commands/admin/TicketSettings.java
@@ -4,6 +4,7 @@ import de.eldoria.shepard.contexts.commands.Command;
 import de.eldoria.shepard.contexts.commands.CommandArg;
 import de.eldoria.shepard.database.queries.TicketData;
 import de.eldoria.shepard.database.types.TicketType;
+import de.eldoria.shepard.wrapper.MessageEventDataWrapper;
 import de.eldoria.shepard.messagehandler.ErrorType;
 import de.eldoria.shepard.messagehandler.MessageSender;
 import de.eldoria.shepard.util.Verifier;
@@ -11,7 +12,6 @@ import net.dv8tion.jda.api.entities.Category;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.entities.TextChannel;
-import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -58,10 +58,10 @@ public class TicketSettings extends Command {
     }
 
     @Override
-    protected void internalExecute(String label, String[] args, MessageReceivedEvent receivedEvent) {
+    protected void internalExecute(String label, String[] args, MessageEventDataWrapper messageContext) {
         String cmd = args[0];
         String type = args[1];
-        List<TicketType> tickets = TicketData.getTypes(receivedEvent.getGuild(), receivedEvent);
+        List<TicketType> tickets = TicketData.getTypes(messageContext.getGuild(), messageContext);
         TicketType scopeTicket = null;
         for (TicketType ticket : tickets) {
             if (ticket.getKeyword().equalsIgnoreCase(type)) {
@@ -74,44 +74,44 @@ public class TicketSettings extends Command {
 
 
         if ((cmd.equalsIgnoreCase("createType") || cmd.equalsIgnoreCase("ct")) && scopeTicket != null) {
-            createType(args, receivedEvent, type);
+            createType(args, messageContext, type);
             return;
         } else if ((cmd.equalsIgnoreCase("createType") || cmd.equalsIgnoreCase("ct")) && scopeTicket == null) {
-            MessageSender.sendSimpleError(ErrorType.TYPE_ALREADY_DEFINED, receivedEvent.getChannel());
+            MessageSender.sendSimpleError(ErrorType.TYPE_ALREADY_DEFINED, messageContext.getChannel());
             return;
         }
 
         if (scopeTicket == null) {
-            MessageSender.sendSimpleError(ErrorType.TYPE_NOT_FOUND, receivedEvent.getChannel());
+            MessageSender.sendSimpleError(ErrorType.TYPE_NOT_FOUND, messageContext.getChannel());
             return;
         }
 
         if (cmd.equalsIgnoreCase("removeType") || cmd.equalsIgnoreCase("rt")) {
-            removeType(args, receivedEvent, scopeTicket);
+            removeType(args, messageContext, scopeTicket);
             return;
         }
 
         if (cmd.equalsIgnoreCase("setOwnerRoles") || cmd.equalsIgnoreCase("sor")
                 || cmd.equalsIgnoreCase("setSupportRoles") || cmd.equalsIgnoreCase("ssr")) {
-            setRoles(args, receivedEvent, cmd, scopeTicket);
+            setRoles(args, messageContext, cmd, scopeTicket);
             return;
         }
 
         if (cmd.equalsIgnoreCase("setChannelCategory") || cmd.equalsIgnoreCase("scc")) {
-            setChannelCategory(args, receivedEvent, scopeTicket);
+            setChannelCategory(args, messageContext, scopeTicket);
             return;
         }
 
         if (cmd.equalsIgnoreCase("setCreationMessage") || cmd.equalsIgnoreCase("scm")) {
-            setCreationMessage(args, receivedEvent, scopeTicket);
+            setCreationMessage(args, messageContext, scopeTicket);
             return;
         }
 
-        MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, receivedEvent.getChannel());
-        sendCommandUsage(receivedEvent.getChannel());
+        MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, messageContext.getChannel());
+        sendCommandUsage(messageContext.getChannel());
     }
 
-    private void setCreationMessage(String[] args, MessageReceivedEvent receivedEvent, TicketType scopeTicket) {
+    private void setCreationMessage(String[] args, MessageEventDataWrapper receivedEvent, TicketType scopeTicket) {
         if (args.length < 3) {
             MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, receivedEvent.getChannel());
             sendCommandUsage(receivedEvent.getChannel());
@@ -120,14 +120,14 @@ public class TicketSettings extends Command {
 
         String message = String.join(" ", Arrays.copyOfRange(args, 2, args.length));
 
-        TicketData.addType(receivedEvent.getGuild(), null, message,
-                scopeTicket.getKeyword(), receivedEvent);
-        MessageSender.sendSimpleTextBox("Set creation text for ticket type " + scopeTicket.getKeyword() + " to:",
-                message, receivedEvent.getChannel());
-
+        if (TicketData.setCreationMessage(receivedEvent.getGuild(), scopeTicket.getKeyword(), message,
+                receivedEvent)) {
+            MessageSender.sendSimpleTextBox("Set creation text for ticket type " + scopeTicket.getKeyword() + " to:",
+                    message, receivedEvent.getChannel());
+        }
     }
 
-    private void setChannelCategory(String[] args, MessageReceivedEvent receivedEvent, TicketType scopeTicket) {
+    private void setChannelCategory(String[] args, MessageEventDataWrapper receivedEvent, TicketType scopeTicket) {
         if (args.length != 3) {
             MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, receivedEvent.getChannel());
             sendCommandUsage(receivedEvent.getChannel());
@@ -141,13 +141,14 @@ public class TicketSettings extends Command {
             return;
         }
 
-        TicketData.addType(receivedEvent.getGuild(), category, null,
-                scopeTicket.getKeyword(), receivedEvent);
-        MessageSender.sendMessage("Set channel category for ticket type \"" + scopeTicket.getKeyword()
-                + "\" to " + category.getName(), receivedEvent.getChannel());
+        if (TicketData.addType(receivedEvent.getGuild(), category, null,
+                scopeTicket.getKeyword(), receivedEvent)) {
+            MessageSender.sendMessage("Set channel category for ticket type \"" + scopeTicket.getKeyword()
+                    + "\" to " + category.getName(), receivedEvent.getChannel());
+        }
     }
 
-    private void setRoles(String[] args, MessageReceivedEvent receivedEvent, String cmd, TicketType scopeTicket) {
+    private void setRoles(String[] args, MessageEventDataWrapper receivedEvent, String cmd, TicketType scopeTicket) {
         if (args.length < 3) {
             MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, receivedEvent.getChannel());
             sendCommandUsage(receivedEvent.getChannel());
@@ -160,28 +161,29 @@ public class TicketSettings extends Command {
         validRoles.forEach(role -> roleMentions.add(role.getAsMention()));
 
         if (cmd.equalsIgnoreCase("setOwnerRoles") || cmd.equalsIgnoreCase("sor")) {
-            TicketData.setTypeOwnerRoles(receivedEvent.getGuild(), scopeTicket.getKeyword(), validRoles, receivedEvent);
+            if (TicketData.setTypeOwnerRoles(receivedEvent.getGuild(), scopeTicket.getKeyword(),
+                    validRoles, receivedEvent)) {
 
-            MessageSender.sendSimpleTextBox("Set the following roles as owner roles for ticket "
-                            + scopeTicket.getKeyword(), String.join(lineSeparator() + "", roleMentions),
-                    receivedEvent.getChannel());
-        } else {
-            TicketData.setTypeSupportRoles(receivedEvent.getGuild(), scopeTicket.getKeyword(),
-                    validRoles, receivedEvent);
+                MessageSender.sendSimpleTextBox("Set the following roles as owner roles for ticket "
+                                + scopeTicket.getKeyword(), String.join(lineSeparator() + "", roleMentions),
+                        receivedEvent.getChannel());
+            }
 
+        } else if (TicketData.setTypeSupportRoles(receivedEvent.getGuild(), scopeTicket.getKeyword(),
+                validRoles, receivedEvent)) {
             MessageSender.sendSimpleTextBox("Set the following roles as owner roles for ticket "
                             + scopeTicket.getKeyword(), String.join(lineSeparator() + "", roleMentions),
                     receivedEvent.getChannel());
         }
+
     }
 
-    private void removeType(String[] args, MessageReceivedEvent receivedEvent, TicketType scopeTicket) {
+    private void removeType(String[] args, MessageEventDataWrapper receivedEvent, TicketType scopeTicket) {
         if (args.length != 2) {
             MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, receivedEvent.getChannel());
             sendCommandUsage(receivedEvent.getChannel());
             return;
         }
-
         List<String> channelIdsByType = TicketData.getChannelIdsByType(receivedEvent.getGuild(),
                 scopeTicket.getKeyword(), receivedEvent);
 
@@ -191,10 +193,13 @@ public class TicketSettings extends Command {
         List<String> typeOwnerRoles = TicketData.getTypeOwnerRoles(receivedEvent.getGuild(),
                 scopeTicket.getKeyword(), receivedEvent);
 
+
         Set<Member> members = new HashSet<>();
 
         for (TextChannel channel : validTextChannels) {
+
             String channelOwnerId = TicketData.getChannelOwnerId(receivedEvent.getGuild(), channel, receivedEvent);
+
             if (channelOwnerId == null) continue;
             Member memberById = receivedEvent.getGuild().getMemberById(channelOwnerId);
             if (memberById == null) continue;
@@ -204,18 +209,17 @@ public class TicketSettings extends Command {
             TicketHelper.removeAndUpdateTicketRoles(receivedEvent, member, typeOwnerRoles);
         }
 
-        TicketData.removeTypeByKeyword(receivedEvent.getGuild(), scopeTicket.getKeyword(), receivedEvent);
-
-        for (TextChannel channel : validTextChannels) {
-            channel.delete().queue();
+        if (TicketData.removeTypeByKeyword(receivedEvent.getGuild(), scopeTicket.getKeyword(), receivedEvent)) {
+            for (TextChannel channel : validTextChannels) {
+                channel.delete().queue();
+            }
+            MessageSender.sendMessage("Removed ticket type **" + scopeTicket.getKeyword()
+                            + "** and all channels of this type!",
+                    receivedEvent.getChannel());
         }
-
-        MessageSender.sendMessage("Removed ticket type **" + scopeTicket.getKeyword()
-                        + "** and all channels of this type!",
-                receivedEvent.getChannel());
     }
 
-    private void createType(String[] args, MessageReceivedEvent receivedEvent, String type) {
+    private void createType(String[] args, MessageEventDataWrapper receivedEvent, String type) {
         if (args.length != 3) {
             MessageSender.sendSimpleError(ErrorType.INVALID_ARGUMENT, receivedEvent.getChannel());
             sendCommandUsage(receivedEvent.getChannel());
@@ -230,8 +234,9 @@ public class TicketSettings extends Command {
             return;
         }
 
-        TicketData.addType(receivedEvent.getGuild(), category, "", type, receivedEvent);
-
-        MessageSender.sendMessage("Created ticket type: **" + type.toLowerCase() + "**", receivedEvent.getChannel());
+        if (TicketData.addType(receivedEvent.getGuild(), category, "", type, receivedEvent)) {
+            MessageSender.sendMessage("Created ticket type: **"
+                    + type.toLowerCase() + "**", receivedEvent.getChannel());
+        }
     }
 }
