+++ src/main/java/de/eldoria/shepard/listener/TicketCleanupListener.java
@@ -17,11 +17,13 @@ public class TicketCleanupListener extends ListenerAdapter {
     public void onGuildMemberLeave(@Nonnull GuildMemberLeaveEvent event) {
         List<String> channelIds = getChannelIdsByOwner(event.getGuild(), event.getUser(), null);
 
+
         List<TextChannel> validTextChannels = Verifier.getValidTextChannels(event.getGuild(), channelIds);
         for (TextChannel channel : validTextChannels) {
-            removeChannel(event.getGuild(), channel, null);
-            channel.delete().queue();
-        }
+            if (removeChannel(event.getGuild(), channel, null)) {
+                channel.delete().queue();
+            }
 
+        }
     }
 }
